Add FilterRule type and use it in filters form
